Modified positions of enemies and items to spawn/drop in the correct spot. also fixed wall master bug where they would turn into black squares
